feat: complete modal
