+++ src/components/Modal/CardModal/CardModal.module.css
@@ -1,0 +1,78 @@
+.content {
+  position: relative;
+
+  width: 100%;
+  height: 100%;
+  min-height: 200px;
+
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 20px;
+
+  padding: 30px 20px 20px 20px;
+  border-radius: 20px;
+
+  background-color: #991788;
+
+  font-size: 26px;
+  color: white;
+}
+
+.close {
+  position: absolute;
+  top: 5px;
+  right: 10px;
+
+  width: 30px;
+  height: 30px;
+
+  background-image: url('../../../assets/svg/cross.svg');
+}
+
+.top,
+.bottom {
+  display: flex;
+  height: 100%;
+}
+
+.top {
+  align-items: center;
+}
+
+.image {
+  height: 100%;
+  border-radius: 10px;
+}
+
+.bottom {
+  flex-direction: column;
+  flex: 1;
+  align-items: center;
+}
+
+.name {
+  margin: 0 0 10px 0;
+  color: gold;
+  font-size: 32px;
+  text-align: center;
+}
+
+.info {
+  display: flex;
+  flex-direction: column;
+  gap: 7px;
+}
+
+.infobubble {
+  padding: 5px;
+  border: 1px solid gold;
+  background-color: #e405c6;
+  border-radius: 10px;
+}
+
+@media (max-width: 650px) {
+  .content {
+    flex-direction: column;
+  }
+}
