+++ src/components/Modal/Modal.module.css
@@ -2,15 +2,26 @@
   position: fixed;
   top: 0;
   bottom: 0;
+
   width: 100%;
   height: 100%;
+
   display: flex;
   justify-content: center;
   align-items: center;
+
+  background: rgba(0, 0, 0, 0.5);
 }
 
 .modal {
-  width: 50%;
-  height: 50%;
+  width: 700px;
+  min-height: 200px;
   background-color: white;
+  border-radius: 20px;
+}
+
+@media (max-width: 800px) {
+  .modal {
+    width: auto;
+  }
 }
