+++ src/pages/MainPage/MainPage.tsx
@@ -5,7 +5,8 @@ import CardsList from '../../components/CardsList/CardsList';
 import { joinClassNames } from '../../utils/utils';
 import { IMainCard } from 'types/interfaces';
 import useFetchByQuery from '../../hooks/useFetchByQuery';
-import CardModal from '../../components/CardModal/CardModal';
+import CardModal from '../../components/Modal/CardModal/CardModal';
+import Modal from '../../components/Modal/Modal';
 import { apiLink } from '../../config/config';
 
 const MainPage = () => {
@@ -32,7 +33,11 @@ const MainPage = () => {
     <div className={joinClassNames(styles.main, 'page')}>
       <Search query={query} setQuery={setQuery} />
       {content}
-      {isModalOpen && <CardModal setModalOpen={setModalOpen} cardId={cardId} />}
+      {isModalOpen && (
+        <Modal>
+          <CardModal setModalOpen={setModalOpen} cardId={cardId} />
+        </Modal>
+      )}
     </div>
   );
 };
